update from v2.2.1 to v2.2.2

+++ Sample/ThreadPool/Eterfree.cpp
@@ -1,14 +1,5 @@
-﻿#define ETERFREE
-#define BOOST
-
-//#define FILE_STREAM
-//#define FILE_SYSTEM
-
-#if defined ETERFREE
+﻿#include "Common.h"
 #include "ThreadPool.h"
-#elif defined BOOST
-#include <boost/threadpool.hpp>
-#endif
 
 #include <chrono>
 #include <cstdlib>
@@ -17,26 +8,9 @@
 #include <atomic>
 #include <thread>
 
-#ifdef FILE_STREAM
-#include <fstream>
-
-#ifdef FILE_SYSTEM
-#include <filesystem>
-#endif // FILE_SYSTEM
-#endif // FILE_STREAM
-
-#ifdef _WIN32
-#include <Windows.h>
-#pragma comment(lib, "WinMM.Lib")
-#endif // _WIN32
-
-#if defined ETERFREE
 using ThreadPool = Eterfree::ThreadPool;
-#elif defined BOOST
-using ThreadPool = boost::threadpool::thread_pool<>;
-#endif
 
-static constexpr auto SLEEP_TIME = std::chrono::milliseconds(1);
+static constexpr std::chrono::nanoseconds::rep SLEEP_TIME = 1000000;
 
 static std::atomic_ulong counter = 0;
 
@@ -45,29 +19,12 @@ static void task()
 	for (volatile auto index = 0UL; \
 		index < 10000UL; ++index);
 
-#ifdef _WIN32
-	constexpr UINT PERIOD = 1;
-
-	auto result = ::timeBeginPeriod(PERIOD);
-	if (result != TIMERR_NOERROR)
-		std::cerr << "timeBeginPeriod error " \
-		<< result << std::endl;
-#endif // _WIN32
-
-	std::this_thread::sleep_for(SLEEP_TIME);
-
-#ifdef _WIN32
-	result = ::timeEndPeriod(PERIOD);
-	if (result != TIMERR_NOERROR)
-		std::cerr << "timeEndPeriod error " \
-		<< result << std::endl;
-#endif // _WIN32
+	sleepFor(SLEEP_TIME);
 
 	counter.fetch_add(1, \
 		std::memory_order_relaxed);
 }
 
-#if defined ETERFREE
 static void execute(ThreadPool& _threadPool)
 {
 	auto proxy = _threadPool.getProxy();
@@ -87,54 +44,28 @@ static void terminate(ThreadPool&& _threadPool)
 	(void)threadPool;
 }
 
-#elif defined BOOST
-static auto getConcurrency() noexcept
-{
-	auto concurrency = std::thread::hardware_concurrency();
-	return concurrency > 0 ? \
-		concurrency : static_cast<decltype(concurrency)>(1);
-}
-
-static void execute(ThreadPool& _threadPool)
-{
-	for (auto index = 0UL; index < 50000UL; ++index)
-		_threadPool.schedule(task);
-
-	for (auto index = 0UL; index < 50000UL; ++index)
-		_threadPool.schedule(task);
-}
-
-static void terminate(ThreadPool&& _threadPool)
-{
-	auto threadPool(std::forward<ThreadPool>(_threadPool));
-	(void)threadPool;
-}
-#endif
-
 int main()
 {
 	using namespace std::chrono;
+
 	using std::cout, std::endl;
 
 	constexpr auto load = []() noexcept
 	{ return counter.load(std::memory_order_relaxed); };
 
 #ifdef FILE_STREAM
-	constexpr auto FILE = "ThreadPool.txt";
+	constexpr auto FILE = "Eterfree.txt";
 
 #ifdef FILE_SYSTEM
-	std::filesystem::remove(FILE);
+	if (std::filesystem::exists(FILE))
+		std::filesystem::remove(FILE);
 #endif // FILE_SYSTEM
 
 	std::ofstream ofs(FILE, std::ios::app);
 	auto os = cout.rdbuf(ofs.rdbuf());
 #endif // FILE_STREAM
 
-#if defined ETERFREE
 	ThreadPool threadPool;
-#elif defined BOOST
-	ThreadPool threadPool(getConcurrency());
-#endif
 
 	auto begin = system_clock::now();
 	execute(threadPool);
@@ -156,3 +87,5 @@ int main()
 	cout << "任务总数：" << load() << endl;
 	return EXIT_SUCCESS;
 }
+
+#include "Common.cpp"
